producer-api/requirements.txt
Flask==3.1.2
kafka-python==2.3.0
